libc/fildes: Declare struct file_tx_ops next to struct file_tx
Struct file_tx_ops is the vtable for struct file_tx. Declare it next
to the latter.
