+++ modules/libc/src/fildes/file_tx.h
@@ -19,9 +19,14 @@
 
 #pragma once
 
+#include "picotm/picotm-libc.h"
 #include "picotm/picotm-lib-ptr.h"
 #include "picotm/picotm-lib-ref.h"
 #include "picotm/picotm-lib-slist.h"
+
+#include <stdbool.h>
+#include <sys/socket.h>
+#include <sys/types.h>
 
 /**
  * \cond impl || libc_impl || libc_impl_fd
@@ -31,8 +36,193 @@
  * \endcond
  */
 
-struct picotm_error;
 struct file;
+struct file_tx;
+struct picotm_error;
+struct stat;
+
+union fcntl_arg;
+
+/**
+ * \brief File operations.
+ */
+struct file_tx_ops {
+
+    enum picotm_libc_file_type type;
+
+    /*
+     * File handling
+     */
+
+    void (*prepare)(struct file_tx*, struct file*, void*, struct picotm_error*);
+    void (*release)(struct file_tx*);
+
+    /*
+     * Module interfaces
+     */
+
+    void (*finish)(struct file_tx*);
+
+    /*
+     * accept()
+     */
+
+    int (*accept_exec)(struct file_tx*,
+                       int, struct sockaddr*, socklen_t*,
+                       bool, int*, struct picotm_error*);
+    void (*accept_apply)(struct file_tx*, int, int, struct picotm_error*);
+    void (*accept_undo)(struct file_tx*, int, int, struct picotm_error*);
+
+    /*
+     * bind()
+     */
+
+    int (*bind_exec)(struct file_tx*,
+                     int, const struct sockaddr*, socklen_t,
+                     bool, int*, struct picotm_error*);
+    void (*bind_apply)(struct file_tx*, int, int, struct picotm_error*);
+    void (*bind_undo)(struct file_tx*, int, int, struct picotm_error*);
+
+    /*
+     * connect()
+     */
+
+    int (*connect_exec)(struct file_tx*,
+                        int, const struct sockaddr*, socklen_t,
+                        bool, int*, struct picotm_error*);
+    void (*connect_apply)(struct file_tx*, int, int, struct picotm_error*);
+    void (*connect_undo)(struct file_tx*, int, int, struct picotm_error*);
+
+    /*
+     * fchmod()
+     */
+
+    int (*fchmod_exec)(struct file_tx*,
+                       int, mode_t,
+                       bool, int*, struct picotm_error*);
+    void (*fchmod_apply)(struct file_tx*, int, int, struct picotm_error*);
+    void (*fchmod_undo)(struct file_tx*, int, int, struct picotm_error*);
+
+    /*
+     * fcntl()
+     */
+
+    int (*fcntl_exec)(struct file_tx*,
+                      int, int, union fcntl_arg*,
+                      bool, int*, struct picotm_error*);
+    void (*fcntl_apply)(struct file_tx*, int, int, struct picotm_error*);
+    void (*fcntl_undo)(struct file_tx*, int, int, struct picotm_error*);
+
+    /*
+     * fstat()
+     */
+
+    int (*fstat_exec)(struct file_tx*,
+                      int, struct stat*,
+                      bool, int*, struct picotm_error*);
+    void (*fstat_apply)(struct file_tx*, int, int, struct picotm_error*);
+    void (*fstat_undo)(struct file_tx*, int, int, struct picotm_error*);
+
+    /*
+     * fsync()
+     */
+
+    int (*fsync_exec)(struct file_tx*,
+                      int,
+                      bool, int*, struct picotm_error*);
+    void (*fsync_apply)(struct file_tx*, int, int, struct picotm_error*);
+    void (*fsync_undo)(struct file_tx*, int, int, struct picotm_error*);
+
+    /*
+     * listen()
+     */
+
+    int (*listen_exec)(struct file_tx*,
+                       int, int,
+                       bool, int*, struct picotm_error*);
+    void (*listen_apply)(struct file_tx*, int, int, struct picotm_error*);
+    void (*listen_undo)(struct file_tx*, int, int, struct picotm_error*);
+
+    /*
+     * lseek()
+     */
+
+    off_t (*lseek_exec)(struct file_tx*,
+                        int, off_t, int,
+                        bool, int*, struct picotm_error*);
+    void (*lseek_apply)(struct file_tx*, int, int, struct picotm_error*);
+    void (*lseek_undo)(struct file_tx*, int, int, struct picotm_error*);
+
+    /*
+     * pread()
+     */
+
+    ssize_t (*pread_exec)(struct file_tx*,
+                          int, void*, size_t, off_t,
+                          bool, int*, struct picotm_error*);
+    void (*pread_apply)(struct file_tx*, int, int, struct picotm_error*);
+    void (*pread_undo)(struct file_tx*, int, int, struct picotm_error*);
+
+    /*
+     * pwrite()
+     */
+
+    ssize_t (*pwrite_exec)(struct file_tx*,
+                           int, const void*, size_t, off_t,
+                           bool, int*, struct picotm_error*);
+    void (*pwrite_apply)(struct file_tx*, int, int, struct picotm_error*);
+    void (*pwrite_undo)(struct file_tx*, int, int, struct picotm_error*);
+
+    /*
+     * read()
+     */
+
+    ssize_t (*read_exec)(struct file_tx*,
+                         int, void *buf, size_t,
+                         bool, int*, struct picotm_error*);
+    void (*read_apply)(struct file_tx*, int, int, struct picotm_error*);
+    void (*read_undo)(struct file_tx*, int, int, struct picotm_error*);
+
+    /*
+     * recv()
+     */
+
+    ssize_t (*recv_exec)(struct file_tx*,
+                         int, void*, size_t, int,
+                         bool, int*, struct picotm_error*);
+    void (*recv_apply)(struct file_tx*, int, int, struct picotm_error*);
+    void (*recv_undo)(struct file_tx*, int, int, struct picotm_error*);
+
+    /*
+     * send()
+     */
+
+    ssize_t (*send_exec)(struct file_tx*,
+                         int, const void*, size_t, int,
+                         bool, int*, struct picotm_error*);
+    void (*send_apply)(struct file_tx*, int, int, struct picotm_error*);
+    void (*send_undo)(struct file_tx*, int, int, struct picotm_error*);
+
+    /*
+     * shutdown()
+     */
+
+    int (*shutdown_exec)(struct file_tx*,
+                         int, int,
+                         bool, int*, struct picotm_error*);
+    void (*shutdown_apply)(struct file_tx*, int, int, struct picotm_error*);
+    void (*shutdown_undo)(struct file_tx*, int, int, struct picotm_error*);
+
+    /*
+     * write()
+     */
+
+    ssize_t (*write_exec)(struct file_tx*,
+                          int, const void*, size_t,
+                          bool, int*, struct picotm_error*);
+    void (*write_apply)(struct file_tx*, int, int, struct picotm_error*);
+    void (*write_undo)(struct file_tx*, int, int, struct picotm_error*);
+};
 
 /**
  * Holds transaction-local state for a file.
